Fix bugs in unit cast evaluation
One of the bugs was because we were using `value_binop` directly, when
in fact we should have been substituting the units manually.

The other bug involved an oversight of the fact that, when converting
units, you must multiply to convert in one direction and divide to
convert in the other. So, one of the sides actually needed to be divided
instead of multiplied.