Don't leak AST memory
The leak is a leftover from when the EvalContext would store pointers to
AST nodes allocated in the parser; however, now, it evaluates
expressions in e.g. variable/unit definitions greedily, so storing (a
pointer to) the expression itself is no longer necessary, and the memory
can be cleaned up.